# docs/swagger.yaml
basePath: /api
definitions:
...
  models.PurchaseOrder:
    type: object
  models.Recipe:
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@b-resto.com
    name: Soporte API
  description: API RESTful para sistema POS de restaurantes
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: B-Resto API
  version: "1.0"
paths:
...
  /purchase-orders:
    get:
      consumes:
      - application/json
      description: Obtiene lista de todas las órdenes de compra
      parameters:
      - description: Filtrar por estado
        enum:
        - draft
        - sent
        - received
        - cancelled
        in: query
        name: status
        type: string
      - description: Filtrar por proveedor
        in: query
        name: partner_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: array de purchase orders'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: mensaje'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar órdenes de compra
      tags:
      - purchase-orders
    post:
      consumes:
      - application/json
      description: Crea una nueva orden de compra
      parameters:
      - description: Datos de la orden
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.PurchaseOrder'
      produces:
      - application/json
      responses:
        "201":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Crear orden de compra
      tags:
      - purchase-orders
  /purchase-orders/{id}:
    get:
      consumes:
      - application/json
      description: Obtiene una orden de compra por ID con todos sus detalles
      parameters:
      - description: ID de la orden de compra
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: purchase order completa'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Purchase order not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener orden de compra
      tags:
      - purchase-orders
    put:
      consumes:
      - application/json
      description: Actualiza los datos de una orden de compra existente
      parameters:
      - description: ID de la orden
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.PurchaseOrder'
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Purchase order not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Actualizar orden de compra
      tags:
      - purchase-orders
  /purchase-orders/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cambia el estado de la orden a cancelled
      parameters:
      - description: ID de la orden
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Purchase order not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Cancelar orden de compra
      tags:
      - purchase-orders
  /purchase-orders/{id}/receive:
    patch:
      consumes:
      - application/json
      description: Cambia el estado de la orden a received (actualiza inventario)
      parameters:
      - description: ID de la orden
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Purchase order not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Recibir orden de compra
      tags:
      - purchase-orders
  /purchase-orders/{id}/send:
    patch:
      consumes:
      - application/json
      description: Cambia el estado de la orden a sent
      parameters:
      - description: ID de la orden
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Purchase order not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Enviar orden de compra
      tags:
      - purchase-orders
  /recipes:
    get:
      consumes:
      - application/json
      description: Obtiene lista de todas las recetas de productos
      parameters:
      - description: Filtrar por producto
        in: query
        name: product_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: array de recipes'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: mensaje'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar recetas
      tags:
      - recipes
    post:
      consumes:
      - application/json
      description: Crea una nueva receta de producto con sus ingredientes
      parameters:
      - description: Datos de la receta
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "201":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Crear receta
      tags:
      - recipes
  /recipes/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una receta (soft delete)
      parameters:
      - description: ID de la receta
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Recipe deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Recipe not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Eliminar receta
      tags:
      - recipes
    get:
      consumes:
      - application/json
      description: Obtiene una receta por ID con todos sus ingredientes
      parameters:
      - description: ID de la receta
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: recipe con items'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Recipe not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener receta
      tags:
      - recipes
    put:
      consumes:
      - application/json
      description: Actualiza los datos de una receta existente
      parameters:
      - description: ID de la receta
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Recipe not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Actualizar receta
      tags:
      - recipes
...
securityDefinitions:
  Bearer:
    description: Escribir "Bearer {token}" para autenticación JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"